Add 'estimate tasks' so you can process your inbox
The flow is, you say 'estimate tasks'

Synergy then says:

    Task: <name> - <link>

And you then say:

    estimate /e .5h-1h

You can use any of the / specials that can be passed in
when creating a new task (or, you'll be able to once update
task gets that functionality).

Ideally this would give you the description and what project
owner is and automatically assign it back to the correct user
if it can.